Implement orderbook with new styles
