+++ src/pages/LendPage/components/ExpandableCardContent/ExpandableCardContent.module.less
@@ -23,6 +23,5 @@
   border: var(--border-less-primary);
   border-right: 0;
   border-bottom: 0;
-  overflow-x: hidden;
   width: 100%;
 }
