+++ src/pages/LendPage/components/Offer/Offer.module.less
@@ -1,10 +1,10 @@
 .listItem {
   position: relative;
-  padding: 8px 48px 8px 8px;
+  padding: 4px 64px 4px 24px;
   background: var(--bg-action);
   display: flex;
   align-items: center;
-  min-height: 48px;
+  min-height: 32px;
   width: 100%;
 
   &:not(:last-child) {
@@ -18,7 +18,6 @@
 
   font: var(--table-header-sm);
   color: var(--pure-black);
-  text-align: center;
   text-transform: uppercase;
 
   &::after,
@@ -66,18 +65,26 @@
   align-items: center;
   justify-content: space-between;
   width: 100%;
+
+  p {
+    width: calc(100% / 3);
+
+    &:nth-child(2) {
+      text-align: center;
+    }
+
+    &:last-child {
+      text-align: right;
+    }
+  }
 }
 
 .value {
   color: var(--content-primary);
-  display: flex;
-  gap: 4px;
   font: var(--body-text-md);
-
   line-height: 24px;
   overflow: auto;
   white-space: nowrap;
-  max-width: 200px;
 }
 
 .editButton {
